Fixed  delete and update user

+++ kabina_project_api/src/main/java/Kabina/Service/UserService.java
@@ -6,6 +6,6 @@ public interface UserService {
 	Iterable<User> findAll();
 	public User addUser(User usr);
 	public User updateUser(User usr);
-	public void deleteUser (Integer Id);
+	public void deleteUser (Long Id);
 	int checkUserExist(String userName);
 }
